requirements.txt
langchain-openai>=1.1.10
langchain-community>=0.4.1
langchain-core>=1.2.0
langchain>=1.2.0
streamlit>=1.28.0
wikipedia>=1.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
stop-words>=2018.7.23
